fix: split terminal host lifecycle helpers under max-lines ratchet
Move probe/shutdown helpers into a part file so terminal_host_client.dart stays within its baseline allowance.
